requirements.txt
tweepy==3.7.0
kafka-python==1.4.5